refactor: move button-specific styles to button component
